requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
flask-cors==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
urllib3==2.0.5
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
click==8.1.7
blinker==1.6.2
gunicorn==21.2.0
